fix(testing): add RecordingWidget to oriterm_ui::testing, migrate FocusFallthroughProbe
Phase 4 implementation for BUG-07-023. Adds RecordingWidget — a shared
test-only Widget that records every InputEvent via on_input, exposes it
through a cloneable RecordedEvents handle (count_keydowns, last_event,
all, len, is_empty). Backed by Rc<RefCell<Vec<InputEvent>>> since Widget
has no Send bound.

Migrates 3 overlay/tests.rs tests away from the bespoke FocusFallthroughProbe
and its Arc<AtomicUsize> plumbing, deleting the probe struct entirely.
New home per .claude/rules/tests.md §Widget Harness Testing.

Collateral fixes (broken-window policy):
- oriterm_mux/src/testing/mod.rs: clippy::extend_with_drain -> append
- oriterm/src/font/discovery/unix.rs: remove redundant `as u32` cast
- oriterm/src/app/redraw/mod.rs: add #[expect] for too_many_arguments with reason
- 6 prose-lint bug-id violations stripped from doc comments and section banners

+++ oriterm/src/font/discovery/unix.rs
@@ -100,8 +100,8 @@ pub(super) fn enumerate_mono_families_from_roots(roots: &[PathBuf]) -> Vec<Famil
 
 /// Parse font file(s) into [`RawFaceInfo`] entries.
 ///
-/// For `.ttc` collections, returns one entry per face (BUG-04-007). For
-/// standalone `.ttf`/`.otf`, returns a single-element `Vec`.
+/// For `.ttc` collections, returns one entry per face. For standalone
+/// `.ttf`/`.otf`, returns a single-element `Vec`.
 #[expect(
     unsafe_code,
     reason = "memmap2::Mmap::map is unsafe by API; font files are read-only system resources"
@@ -125,7 +125,7 @@ fn parse_face_info(path: &Path) -> Vec<RawFaceInfo> {
     let face_pairs: Vec<(skrifa::FontRef<'_>, u32)> = match file_ref {
         FileRef::Font(f) => vec![(f, 0u32)],
         FileRef::Collection(c) => {
-            let count = c.len() as u32;
+            let count = c.len();
             (0..count)
                 .filter_map(|i| c.get(i).ok().map(|f| (f, i)))
                 .collect()
